Initial Commit
This is still a work in progress, I am actively tidying things up, adding documentation, and trimming fat.

There is a known crash when using this plugin in FL Studio on windows, but it works with other DAWs, Operating systems, and it also works in standalone mode.

This was built with JUCE 7.0.12, but it should also work with any version after that.

This also requires the BDSP_Library to function.