Cleaned Up A Few Minor Issues
The crash when loading the plugin in FL Studio on Windows has been fixed with the latest commit to the BDSP_Library.

Removed old commented out code.

Updated uses of the Processor's sample rate in the Editor to the new sample rate value stored in the Editor.

Removed some unnecessary code.

Cleaned up some formatting.